feat: expand zod, codesplit, remove custom components
