--- libs/application/templates/health-insurance/src/fields/CitizenshipField/CitizenshipField.tsx
@@ -1,36 +1,0 @@
-import { useEffect, useState } from 'react'
-import { useFormContext } from 'react-hook-form'
-import { getValueViaPath } from '@island.is/application/core'
-import { FieldBaseProps } from '@island.is/application/types'
-import { Box, Input } from '@island.is/island-ui/core'
-import { ExternalDataNationalRegistry } from '../../utils/types'
-
-export const CitizenshipField = ({ field, application }: FieldBaseProps) => {
-  const { id } = field
-  const citizenship = (
-    getValueViaPath(
-      application.externalData,
-      'nationalRegistry',
-    ) as ExternalDataNationalRegistry
-  )?.data?.citizenship
-
-  const [countryData, setCountryData] = useState<string>(
-    JSON.stringify(citizenship),
-  )
-  const { register } = useFormContext()
-
-  useEffect(() => {
-    setCountryData(JSON.stringify({ ...citizenship }))
-  }, [citizenship])
-
-  return (
-    <Box hidden>
-      <Input
-        id={id}
-        {...register(id)}
-        value={countryData}
-        onChange={() => register}
-      />
-    </Box>
-  )
-}
