--- libs/application/templates/health-insurance/src/fields/ConfirmCheckbox/ConfirmCheckbox.tsx
@@ -1,55 +1,0 @@
-import { Controller, useFormContext } from 'react-hook-form'
-import { formatText, getValueViaPath } from '@island.is/application/core'
-import { FieldBaseProps } from '@island.is/application/types'
-import { Box, Checkbox } from '@island.is/island-ui/core'
-import { useLocale } from '@island.is/localization'
-import { m } from '../../lib/messages/messages'
-
-export const ConfirmCheckbox = ({
-  error,
-  field,
-  application,
-}: FieldBaseProps) => {
-  const { id, disabled } = field
-  const { setValue } = useFormContext()
-  const defaultValue = getValueViaPath(
-    application.answers,
-    id as string,
-    false,
-  ) as boolean
-  const { formatMessage } = useLocale()
-
-  return (
-    <Controller
-      name={id}
-      defaultValue={defaultValue}
-      render={({ field: { onChange, value } }) => {
-        return (
-          <Box
-            border="standard"
-            borderColor="blue200"
-            borderRadius="large"
-            padding={3}
-          >
-            <Checkbox
-              id={id}
-              disabled={disabled}
-              name={`${id}`}
-              label={formatText(
-                m.confirmCorrectInfo,
-                application,
-                formatMessage,
-              )}
-              hasError={!!error}
-              checked={value}
-              onChange={(e) => {
-                onChange(e.target.checked)
-                setValue(id as string, e.target.checked)
-              }}
-            />
-          </Box>
-        )
-      }}
-    />
-  )
-}
