+++ libs/application/templates/health-insurance/src/fields/index.ts
@@ -1,8 +1,5 @@
 export { Review } from './Review'
-export { ConfirmCheckbox } from './ConfirmCheckbox/ConfirmCheckbox'
-export { AgentComment } from './AgentComment/AgentComment'
 export { HiddenDateField } from './HiddenDateField/HiddenDateField'
 export { MissingInfoRemarks } from './MissingInfoRemarks/MissingInfoRemarks'
 export { TextWithTooltip } from './TextWithTooltip/TextWithTooltip'
-export { CitizenshipField } from './CitizenshipField/CitizenshipField'
 export { PrerequisiteSummary } from './PrerequisiteSummary/PrerequisiteSummary'
