+++ libs/application/templates/health-insurance/src/forms/HealthInsuranceForm/aplicantInformationSection.ts
@@ -1,0 +1,9 @@
+import { buildSection } from '@island.is/application/core'
+import { m } from '../../lib/messages/messages'
+import { applicantInformationMultiField } from '@island.is/application/ui-forms'
+
+export const applicantInformationSection = buildSection({
+  id: 'applicantInfoSection',
+  title: m.applicantInfoSection,
+  children: [applicantInformationMultiField()],
+})
