+++ libs/application/templates/health-insurance/src/lib/dataSchema.ts
@@ -1,12 +1,20 @@
 import { applicantInformationSchema } from '@island.is/application/ui-forms'
 import * as z from 'zod'
-import { EmploymentStatus, YesNo } from '../utils/constants'
+import { EmploymentStatus, YES, YesNo } from '../utils/constants'
+
+const FileSchema = z.object({
+  name: z.string(),
+  key: z.string(),
+  url: z.string().optional(),
+})
 
 const formerInsurance = z.object({
   registration: z.nativeEnum(YesNo),
   country: z.string().min(1),
   personalId: z.string().min(1),
   institution: z.string().min(1),
+  confirmationOfResidencyDocument: FileSchema.optional(),
+  entitlement: z.nativeEnum(YesNo).optional(),
   entitlementReason: z.string().optional(),
 })
 
@@ -19,5 +27,6 @@ export const dataSchema = z.object({
   formerInsurance,
   hasAdditionalInfo: z.nativeEnum(YesNo),
   additionalRemarks: z.string().optional(),
-  confirmCorrectInfo: z.boolean().refine((v) => v),
+  additionalFiles: z.array(FileSchema).optional(),
+  confirmCorrectInfo: z.array(z.enum([YES])).length(1),
 })
